Tailwind styling cleanup on mobile, laptop, tablet screens
