+++ components/most-popular-celebs/MostPopularCelebs.tsx
@@ -13,7 +13,7 @@ export default function MostPopularCelebs() {
   const celebs = fetchCelebs();
 
   return (
-    <section className="py-12 w-full md:max-w-5xl lg:max-w-5xl xl:max-w-6xl 2xl:max-w-8xl 3xl:max-w-9xl mx-auto outline outline-2 outline-green-500">
+    <section className="py-12 w-full md:max-w-5xl lg:max-w-5xl xl:max-w-6xl 2xl:max-w-8xl 3xl:max-w-9xl mx-auto ">
       <h2 className="text-3xl font-bold text-center mb-8">
         Most Popular Celebrities
       </h2>
